Batch (1/3): extract shared llm_batch helpers; RCT uses them
Move batch_submit / batch_check_and_collect / _convert_messages_for_anthropic
out of the prod-gated experiments/prompt_rct/run_rct.py into
backend/utils/llm_batch.py (dependency-light: no Flask/DB, lazy SDK imports).
Add apply_batch_key_override (non-mutating OPENAI_API_KEY_BATCH overlay).

Refactor the RCT to import these from the shared module — its duplicate
definitions deleted, get_batch_api_keys routed through the shared override,
unused tempfile import dropped. Single source of truth; no behavior change.

Add unit tests (mocked SDK clients): pure helpers + Anthropic/OpenAI submit
and collect paths. Full suite 363 passed.
